+++ apps/next/src/components/Address/ChainListItem.tsx
@@ -1,7 +1,10 @@
 import {
   Box,
+  getHexAlpha,
   ListItem,
+  ListItemIcon,
   Stack,
+  styled,
   Typography,
   TypographyProps,
 } from '@avalabs/k2-alpine';
@@ -13,7 +16,14 @@ import {
   ReactElement,
 } from 'react';
 import { IconBaseProps } from 'react-icons';
-import { ItemIcon } from './ItemIcon';
+
+export const ItemIcon = styled(ListItemIcon)(({ theme }) => ({
+  minWidth: 36,
+  svg: {
+    borderRadius: '50%',
+    boxShadow: `inset 0 0 10px ${getHexAlpha(theme.palette.primary.main, 30)}`,
+  },
+}));
 
 export type Props = {
   Icon: ComponentType<IconBaseProps> | ReactElement<IconBaseProps>;
