+++ apps/next/src/components/Address/index.ts
@@ -1,0 +1,1 @@
+export * from './AddressItem';
